requirements.txt
google-api-python-client>=2.100.0
anthropic>=0.34.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
